Mmp overview (#4)
* first steps towards mmp overviews using d3.js

* display of MMP data within chart

* minimally functional MMP overview. Has several problems:

- Somehow one of the curves is missing, suspect it to be the first
- Not clear what label refers to what curve
- Too many curves, effects barely visible as it is so crowded

Solutions I was thinking of:
- Show only top 10 and bottom 10 MMPs, aka only the most
  relevant MMPs
- Think about a better way to group the MMPs, currently they
  are alphabetic but should actually be ordered via their
  influence on the target column.

Rational would  be: only MMPs that actually influence the target
should be shown, not the once that are not relevant.

I think the SVG height should be adaptive to the no of datapoints.

* scatterplot with datapoints as well

* very minimal MMP plot

* nicely sized plot

* better kernel and kernel size ✨

* white background

* better margins

* Simpler tooltip 👀 showing on hover

Already quite nice as it shows how to bind hover interactions
with the data so easily.

D3.js is really amazing... 😇

* MMP display of points 🤩