initial draft implementation of logs
There is both a stdout and a error log button in the jobs listing.
Clicking on any of the two reveals the output for it in an
overlay text window.

Internally, we use the Silencer class to redirect stdout/err
to two text files, which are subsequently read by the Job listing
Handler.

This should make it possible to get a live preview of the current
logs, presuming that the flushing frequency is high enough
(at first sight this seems to now be the case).

I actually had to remove the Silencer and directly call
the std_out redirect context handler. Might be worth
finding out why that was necessary