Working feature importance on RDKit descriptors.
As long as at least one model is trained using RDKit descriptors,
a feature importance report is created and stored as a SVG file
on-disk. Next, we can expose it in the app within it's own
feature importance tab.

In addition, the RDKit descriptors provide quite a strong
baseline, so it is also valuable to have them in that regard.