RDKit descriptors (#15)
* added mordred fingerprints, many small fixes

also found out that the computation is unneccessarily slow
just because it wasn't vectorized yet.

Therefore, planning to vectorize the feature computation
next. Then, Mordred Descriptor calculation should be
reasonably fast.

* vectorized atom attributor

* More informative logging and rdkit descriptors

Mordred descriptors are too slow to compute on the mutations.
With RDKit descriptors, it is two orders of magnitude faster,
and also reaching about the same performance...

* feature importance table created

* Working feature importance on RDKit descriptors.

As long as at least one model is trained using RDKit descriptors,
a feature importance report is created and stored as a SVG file
on-disk. Next, we can expose it in the app within it's own
feature importance tab.

In addition, the RDKit descriptors provide quite a strong
baseline, so it is also valuable to have them in that regard.